Move calculation of association schema to helper
