+++ open_alchemy/schemas/helpers/__init__.py
@@ -1,5 +1,6 @@
 """Helper functions for processing the schemas."""
 
+from . import association
 from . import backref
 from . import clean
 from . import iterate
